Propagate bus 'close'
This may be related to the other 2 pending PRs, but seems to be clean and to make sense on all levels - and most importantly, address the root of the issue.

Here's what happens: sometimes, the connection closes. The lib/client Client instance would handle it properly, calling it's onclose, which nulls .ps.

Then all sorts of hell would break loose as various things would try to call castv2 methods, but .ps is null.

The solution - propagate the close up the chain, so that whoever uses castv2 and castv2-client will know the connection has closed and handle it properly.